refactor project links
